Context Switch Timer Interface
A generic interface miosix::ContextSwitchTimer has been added to the
kernel that can play the role of either periodic/aperiodic system timer.
Like other kernel interfaces, there should be provided a proper
implementation for cstimer.

Currently the only available implementation is done for cortexM4_stm32f
which is based on TIM2 timer. Although the implementation is an
aperiodic timer, it is set to act like a periodic timer with T=1ms.

Signed-off-by: Sasan Golchin <ahmad.golchin@mail.polimi.it>
Signed-off-by: Terraneo Federico <fede.tft@miosix.org>